Active User Status : Not in Real-time

+++ chatopia/app/api/socket/offline/route.ts
@@ -6,11 +6,21 @@ export async function POST(request : Request){
     try{
         const { email } = await request.json();
         // DELETE the row from active table where id is the email
-        await prisma.active.delete({
+
+        const user = await prisma.user.findUnique({
             where: {
-                id: email
+                email
             }
         });
+
+        if(user){
+            await prisma.active.delete({
+                where: {
+                    email: email
+                }
+            });
+        }
+        
     }
     catch(error: any){
         return new NextResponse('Internal Error', { status: 500 });
